Restore TPE
This restores TPE functionality that was broken when enable_fix_blocks was introduced. .uniform requires 3 arguments, and when the change over was made, these three arguments were not provided to the section that is run when fix blocks is disabled. There will need to probably be another fix for when fix blocks is enabled.