Change Logic for Skip Clip == 1
This actually fixes an oversight I had in the refactor. Due to the way that Bayesian Merger works with models, just returning nothing actually doesn't do what this variable is supposed to do which is keep the clip encoder of Model A. In the event that model A is missing this value, we use the one from model B instead.
Changed if statement, as it may catch too wide a net.