Restore TPE and Fix Merge issues (#47)
* Fix Stage 2 Merge

This fixes the issue in stage 2, where a single key is left behind, thereby breaking the merged model.
This also fixes model name problems due to the lack of using stem on self.model["model_a"] and others like it.

* Restore TPE

This restores TPE functionality that was broken when enable_fix_blocks was introduced. .uniform requires 3 arguments, and when the change over was made, these three arguments were not provided to the section that is run when fix blocks is disabled. There will need to probably be another fix for when fix blocks is enabled.

* Fix Typo

minor typo in this file causes merges to fail during postprocess.

* fix the output of weights in best.log

This wasn't properly declared. Should be fixed now.

* Fix typo

Add variable for real model names, rather than keys

Honestly, refactoring self.model_names to self.model_keys would be better and more clear than the name it currently has.

* Fix Stage 1

Needs more testing. Something is wrong with the merger in general since we aren't matching supermerger, but I don't think it's this logic specifically since we still don't match even on main.

* Properly Refactor skip position/reset clip

This config option has been properly refactored.

Wiki should be updated to reflect that skip_position_ids does not refer to skip clip in WebUI, but corresponds to the clip settings in MBW-GUI.

Removed debug Print functions

Fixed Typo in stage 2

* Change Logic for Skip Clip == 1

This actually fixes an oversight I had in the refactor. Due to the way that Bayesian Merger works with models, just returning nothing actually doesn't do what this variable is supposed to do which is keep the clip encoder of Model A. In the event that model A is missing this value, we use the one from model B instead.
Changed if statement, as it may catch too wide a net.

* Fix logic in Bayes Optimizer to match

This logic matches both the TPE Optimizer and what optimiser.py is expecting from best_bases and best_weights

* Fix formatting in best.log