Implement Cosine Addition and Smooth Add Diff
The implementation for Cosine Addition is not complete nor identical to the one from Supermerger. Currently waiting for clarification on why a certain change was made to the two new methods before replacing the current implementation with them.
Smooth_Add is the same implementation as supermerger, with cupy optimizations built in if a user has that installed and working properly, otherwise it falls back to standard scipy, which is much slower due to being CPU based.