[Feature] Implement TensorSum (#54)
* Implement TensorSum

Taken from the newest commit of supermerger.

* Add to template

* Fix Tensorsum

This fixes a handful of issues which prevented TensorSum from working properly, such as reordering the merge modes in merge_block

* Resolve Style Issue

* Add in wsl related warning for scorer

On WSL, xdg-open will not open the requisite windows applications unless a specialized tool is installed. Adds a warning to warn the user and then tell them where to get the specialized tool needed to properly open the images.

Only display warning once.